Deduplicate common.js and translate email message body
