+++ src/main/webapp/en/index.html
@@ -12,6 +12,7 @@
 
         <script src="https://privacybydesign.foundation/tomcat/irma_api_server/bower_components/jquery/dist/jquery.js" defer></script>
         <script src="https://privacybydesign.foundation/tomcat/irma_api_server/client/irma.js" defer></script>
+        <script src="messages.js" defer></script>
         <script src="common.js" defer></script>
     </head>
     <body>
