+++ src/main/webapp/en/messages.js
@@ -1,0 +1,17 @@
+
+var MESSAGES = {
+    'lang': 'en',
+    'error:email-address-malformed': 'E-mail address is not well-formed',
+    'sending-verification-email': 'A verification e-mail is being sent to %address%…',
+    'sent-verification-email': 'The verification mail has been sent to %address%. Check your inbox to open the verification e-mail. It may take a couple of minutes for the e-mail to arrive. (If you do not see it, check your spam folder too.)',
+    'verifying-email-token': 'E-mail adres is being verified ...',
+    'unknown-problem': 'Unknown problem',
+    'error:invalid-token': 'Link in the e-mail is out-of-date or invalid',
+    'email-failed-to-verify': 'Unknown problem during verification of the e-mail address',
+    'email-add-verified': 'E-mail address verified',
+    'email-add-success': 'E-mail address added',
+    'email-add-cancel': 'Cancelled',
+    'email-add-timeout': 'Sessie time out. Reload this page in order to try again.',
+    'email-add-error': 'Unfortunately, adding this e-mail address to the IRMA app did not succeed.',
+    'return-to-issue-page': '<br><br><a href=\"https://privacybydesign.foundation/issuance/\">Return</a> to attribute issuance.',
+};
