+++ src/main/webapp/nl/messages.js
@@ -1,0 +1,18 @@
+'use strict';
+
+var MESSAGES = {
+    'lang': 'nl',
+    'error:email-address-malformed': 'E-mailadres is onjuist',
+    'sending-verification-email': 'Een verificatie e-mail wordt verzonden aan %address%…',
+    'sent-verification-email': 'De verificatiemail is verzonden aan %address%. Controleer uw inbox voor een verificatie e-mail. Het kan een paar minuten duren voordat de e-mail arriveert. (Controleer zonodig ook uw spam folder.)',
+    'verifying-email-token': 'Het e-mail adres wordt geverifieerd...',
+    'unknown-problem': 'Onbekend probleem',
+    'error:invalid-token': 'De link in de e-mail is verouderd of ongeldig',
+    'email-failed-to-verify': 'Onbekend probleem tijdens het verifiëren van het e-mail adres',
+    'email-add-verified': 'Het e-mail adres geverifieerd',
+    'email-add-success': 'Het e-mail adres is toegevoegd',
+    'email-add-cancel': 'Geannuleerd',
+    'email-add-timeout': 'De sessie is verlopen. Herlaad de pagina om het eventueel opnieuw te proberen.',
+    'email-add-error': 'Het is helaas niet gelukt dit e-mail adres toe te voegen aan de IRMA app.',
+    'return-to-issue-page': '<br><br><a href=\"https://privacybydesign.foundation/uitgifte/\">Terug</a> naar attribuut uitgifte.',
+};
